Auto-heal improvement
Script will now check if the user has said item (read from the config.yaml) in his bag. Previously this was not checked, previously the very first item in the bag would just be used to heal.

Currently the auto heal function does not perform at its best, it sometimes misses and uses the wrong item causing the script to stop.